Display client join/leave messages properly (i.e. they persist when changing current recipient).
Also adds functionality to keep the messages scrollbox scrolled to the bottom.  It additionally ensures that if the user scrolls away from the bottom; it remains where the user kept it.  A button is shown to the user to allow them to easily return to the bottom to follow the message feed.