End-to-end encrypted rooms implementation.
Room can be set to have a symmetric encryption key.  The server notes this, and when a client attempts to join, they send a known message to the room owner who attempts to decrypt it, and. if successful, informs the server that the client has a valid key and is then accepted into the room.

The encryption/decryption code still needs to be abstracted.  The packet-handling switch blocks are also getting oversized; these should be moved into some kind of function table.  There is also notable duplication of some code, which should be moved into methods.