Room ownership transfer.
This fixes issues when owner of encrypted room leaves the server (as he can no longer authorise people to join).  The new system automatically transfers ownership to the next person in the room's list of clients.

For encrypted rooms, if the list of users is empty, the room is simply deleted.

For non-encrypted rooms the room is kept and the new owner is the first person that joins it next.

Room ownership transfers are not yet mentioned in chat feeds.