Server-side message storage + bug fixes.
The server now stores messages that are sent to it, including user messages and user joining, leaving, etc.  For encrypted rooms, the messages are stored in encrypted form and the clients are able to decrypt them when they receive the message list.

Bug where users were unable to join encrypted room after the owner leaves the server was fixed.  This was due to the client not adding the room to their owned room list when the room ownership transfer packet was sent to them.  The owner is required to know that they are the owner of the room so that they can respond to client authorisation requests (and also send requests to the server to delete the room).

Room owner changes are now broadcasted to users in the room.

Password errors are now slightly more specific; internal errors are displayed as so rather than blaming them on incorrect password.

Confirmation dialog added to room deletion menu item (as it's a destructive action).